+++ core/src/main/java/edu/umich/verdict/relation/ApproxRelation.java
@@ -302,67 +302,3 @@ public abstract class ApproxRelation extends Relation {
 
 }
 
-
-class TableNameReplacerInExpr extends ExprModifier {
-
-    final protected Map<TableUniqueName, String> sub;
-
-    public TableNameReplacerInExpr(VerdictContext vc, Map<TableUniqueName, String> sub) {
-        super(vc);
-        this.sub = sub;
-    }
-
-    public Expr call(Expr expr) {
-        if (expr instanceof ColNameExpr) {
-            return replaceColNameExpr((ColNameExpr) expr);
-        } else if (expr instanceof FuncExpr) {
-            return replaceFuncExpr((FuncExpr) expr);
-        } else if (expr instanceof SubqueryExpr) {
-            return replaceSubqueryExpr((SubqueryExpr) expr);
-        } else {
-            return expr;
-        }
-    }
-
-    /**
-     * Replaces if
-     * (1) schema name is omitted in the join condition and the table name matches the table name of the original table
-     * (2) schema name is fully specified in the join condition and both the schema and table names match the original schema and table names.
-     * @param expr
-     * @return
-     */
-    protected Expr replaceColNameExpr(ColNameExpr expr) {
-        if (expr.getTab() == null) {
-            return expr;
-        }
-
-        TableUniqueName old = TableUniqueName.uname(expr.getSchema(), expr.getTab());
-
-        if (old.getSchemaName() == null) {
-            for (Map.Entry<TableUniqueName, String> entry : sub.entrySet()) {
-                TableUniqueName original = entry.getKey();
-                String subAlias = entry.getValue();
-                if (original.getTableName().equals(expr.getTab())) {
-                    return new ColNameExpr(vc, expr.getCol(), subAlias);
-                }
-            }
-        } else {
-            if (sub.containsKey(old)) {
-                String rep = sub.get(old);
-                return new ColNameExpr(vc, expr.getCol(), rep);
-            }
-        }
-        // if nothing matched, then return the original expression.
-        return expr;
-    }
-
-    protected Expr replaceSubqueryExpr(SubqueryExpr expr) {
-        return expr;
-    }
-
-    protected Expr replaceFuncExpr(FuncExpr expr) {
-        FuncExpr e = (FuncExpr) expr;
-        return new FuncExpr(e.getFuncName(), visit(e.getUnaryExpr()));
-    }
-};
-
